Switch from SegmentedList to ArrayList
In master it was removed a couple of days ago,
and since this is such an invasive change,
I'm also doing it on the zig-0.7.x branch now
(which uses zig-0.7.0 to build currently).

Alternatively I could have taken it from
the 'std lib orphanage' repository.
But I've decided not to add that build complexity now,
and keep everything working with an out-of-the-box zig.

(My stance on this might very well change,
as I would like to experiment with things like `zap` for performance,
which also requires some dependency management infrastructure.)